Avoid registering copystructure copiers globally
